+++ lib/apex_charts/options/root_options.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'div_attributes'
+require_relative 'helper_options'
 require_relative 'chart_options'
 require_relative 'stroke_options'
 require_relative 'states_options'
@@ -9,15 +9,11 @@ module ApexCharts::Options
   class RootOptions < ::SmartKv
     optional *%i[
       annotations
-      brushTarget
       chart
-      chartId
       colors
       dataLabels
-      defer
       fill
       grid
-      group
       labels
       legend
       markers
@@ -31,15 +27,10 @@ module ApexCharts::Options
       theme
       title
       tooltip
-      var
       xaxis
-      xtitle
-      xtype
       yaxis
-      ytitle
-      ytype
     ],
-             *DivAttributes.optional_keys,
+             *HelperOptions.optional_keys,
              *ChartOptions.optional_keys,
              *StrokeOptions.optional_keys,
              *StatesOptions.optional_keys