+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Make obsolete previously deprecated local palette creation
+- Update options schema to match ApexCharts.JS version 3.22.0
+- ApexCharts.JS version from 3.21.0 to 3.22.0
 
 ## [0.1.10] - 2020-10-04
 ### Added